Unpack build.tgz once
Instead of unpacking the build.tgz archive, unpack it once into
a volume (zeek_install_data) and mount the contents of that volume
at /zeek/install.

This reduces runtime by not redundantly extracting build.tgz and
also simplifies the setup a bit as we do not need to set ZEEKPATH
anymore (/zeek/install is where Zeek is configured to be installed).