microbenchmarks: Add Intel::insert() microbenchmark
Sometimes, users report memory consumption of the Intel framework with many
indicators as a concern. This PR adds a microbenchmark such that we
start tracking memory usage for 50k and 500k host, subnet and string
indicators over time to capture potential regressions if any of the
involved data structures change. This isn't so much about execution
time, but more about memory usage.